+++ controller/settingscontroller.php
@@ -30,6 +30,7 @@ use OCP\IURLGenerator;
 use OCA\Onlyoffice\AppConfig;
 use OCA\Onlyoffice\Crypt;
 use OCA\Onlyoffice\DocumentService;
+use OCA\Onlyoffice\FileVersions;
 
 /**
  * Settings controller for the administration page
@@ -114,6 +115,7 @@ class SettingsController extends Controller {
             "formats" => $this->config->FormatsSetting(),
             "sameTab" => $this->config->GetSameTab(),
             "preview" => $this->config->GetPreview(),
+            "versionHistory" => $this->config->GetVersionHistory(),
             "limitGroups" => $this->config->GetLimitGroups(),
             "chat" => $this->config->GetCustomizationChat(),
             "compactHeader" => $this->config->GetCustomizationCompactHeader(),
@@ -188,6 +190,7 @@ class SettingsController extends Controller {
      * @param array $editFormats - editable formats array
      * @param bool $sameTab - open in the same tab
      * @param bool $preview - generate preview files
+     * @param bool $versionHistory - keep version history
      * @param array $limitGroups - list of groups
      * @param bool $chat - display chat
      * @param bool $compactHeader - display compact header
@@ -203,6 +206,7 @@ class SettingsController extends Controller {
                                     $editFormats,
                                     $sameTab,
                                     $preview,
+                                    $versionHistory,
                                     $limitGroups,
                                     $chat,
                                     $compactHeader,
@@ -217,6 +221,7 @@ class SettingsController extends Controller {
         $this->config->SetEditableFormats($editFormats);
         $this->config->SetSameTab($sameTab);
         $this->config->SetPreview($preview);
+        $this->config->SetVersionHistory($versionHistory);
         $this->config->SetLimitGroups($limitGroups);
         $this->config->SetCustomizationChat($chat);
         $this->config->SetCustomizationCompactHeader($compactHeader);
@@ -247,6 +252,19 @@ class SettingsController extends Controller {
         }
 
         $this->config->SetWatermarkSettings($settings);
+
+        return [
+            ];
+    }
+
+    /**
+     * Clear all version history
+     *
+     * @return array
+     */
+    public function ClearHistory() {
+
+        FileVersions::clearHistory();
 
         return [
             ];