+++ js/settings.js
@@ -191,6 +191,7 @@
 
             var sameTab = $("#onlyofficeSameTab").is(":checked");
             var preview = $("#onlyofficePreview").is(":checked");
+            var versionHistory = $("#onlyofficeVersionHistory").is(":checked");
 
             var limitGroupsString = $("#onlyofficeGroups").prop("checked") ? $("#onlyofficeLimitGroups").val() : "";
             var limitGroups = limitGroupsString ? limitGroupsString.split("|") : [];
@@ -211,6 +212,7 @@
                     editFormats: editFormats,
                     sameTab: sameTab,
                     preview: preview,
+                    versionHistory: versionHistory,
                     limitGroups: limitGroups,
                     chat: chat,
                     compactHeader: compactHeader,
@@ -279,6 +281,29 @@
             if (code === 13) {
                 $("#onlyofficeAddrSave").click();
             }
+        });
+
+        $("#onlyofficeSecret-show").click(function () {
+            if ($("#onlyofficeSecret").attr("type") == "password") {
+                $("#onlyofficeSecret").attr("type", "text");
+            } else {
+                $("#onlyofficeSecret").attr("type", "password");
+            }
+        });
+
+        $("#onlyofficeClearVersionHistory").click(function () {
+            $(".section-onlyoffice").addClass("icon-loading");
+
+            $.ajax({
+                method: "DELETE",
+                url: OC.generateUrl("apps/" + OCA.Onlyoffice.AppName + "/ajax/settings/history"),
+                success: function onSuccess(response) {
+                    $(".section-onlyoffice").removeClass("icon-loading");
+                    if (response) {
+                        OCP.Toast.success(t(OCA.Onlyoffice.AppName, "All history successfully deleted"));
+                    }
+                }
+            });
         });
     });